+++ src/meltano/core/block/parser.py
@@ -118,8 +118,6 @@ class BlockParser:  # noqa: D101
 
             self.log.debug("found plugin in cli invocation", plugin_name=plugin.name)
 
-        self.log.info("commands", commands=self._commands)
-
     def find_blocks(
         self, offset: int = 0
     ) -> Generator[Union[BlockSet, PluginCommandBlock], None, None]: